# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CurrentReadingDto": {
        "additionalProperties": false,
        "properties": {
          "baromAbsIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "baromRelIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "battOut": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "dailyHighTempF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dailyLowTempF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dailyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "deviceId": {
            "nullable": true,
            "type": "string"
          },
          "deviceName": {
            "nullable": true,
            "type": "string"
          },
          "dewPoint": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dewPointIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "eventRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "feelsLike": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "feelsLikeIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "hourlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "humidity": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "humidityIn": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "lastRain": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "maxDailyGust": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "monthlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "nwsPresentWeather": {
            "nullable": true,
            "type": "string"
          },
          "nwsRawMetar": {
            "nullable": true,
            "type": "string"
          },
          "nwsSkyConditions": {
            "nullable": true,
            "type": "string"
          },
          "nwsTextDescription": {
            "nullable": true,
            "type": "string"
          },
          "omCloudCover": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "omPrecipProbability": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "omPrecipSumIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "omSunrise": {
            "nullable": true,
            "type": "string"
          },
          "omSunset": {
            "nullable": true,
            "type": "string"
          },
          "omUvIndexMax": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "omWeatherDescription": {
            "nullable": true,
            "type": "string"
          },
          "omWindDirDominant": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "omWindGustMax": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "omWindSpeedMax": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "receivedAtUtc": {
            "format": "date-time",
            "type": "string"
          },
          "solarRadiation": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "source": {
            "nullable": true,
            "type": "string"
          },
          "tempF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "tempInF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "timestampUtc": {
            "format": "date-time",
            "type": "string"
          },
          "totalRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "tz": {
            "nullable": true,
            "type": "string"
          },
          "uv": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "weeklyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "windDir": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "windGustMph": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "windSpeedMph": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "yearlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          }
        },
        "required": [
          "deviceId"
        ],
        "type": "object"
      },
      "CustomLayoutItemDto": {
        "additionalProperties": false,
        "properties": {
          "alertsZone": {
            "nullable": true,
            "type": "string"
          },
          "channelStationId": {
            "nullable": true,
            "type": "string"
          },
          "displayMode": {
            "nullable": true,
            "type": "string"
          },
          "icon": {
            "nullable": true,
            "type": "string"
          },
          "iconPosition": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "nullable": true,
            "type": "string"
          },
          "isPaused": {
            "type": "boolean"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/CustomMetricReferenceDto"
            },
            "nullable": true,
            "type": "array"
          },
          "name": {
            "nullable": true,
            "type": "string"
          },
          "position": {
            "nullable": true,
            "type": "string"
          },
          "size": {
            "nullable": true,
            "type": "string"
          },
          "sourceLabels": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "size",
          "type"
        ],
        "type": "object"
      },
      "CustomMetricReferenceDto": {
        "additionalProperties": false,
        "properties": {
          "labelOverride": {
            "nullable": true,
            "type": "string"
          },
          "metricKey": {
            "nullable": true,
            "type": "string"
          },
          "stationId": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "metricKey",
          "stationId"
        ],
        "type": "object"
      },
      "DailyExtremaDto": {
        "additionalProperties": false,
        "properties": {
          "dailyHighTempF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dailyHighTempInF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dailyLowTempF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dailyLowTempInF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dateUtc": {
            "format": "date-time",
            "type": "string"
          },
          "deviceId": {
            "nullable": true,
            "type": "string"
          },
          "deviceName": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "dateUtc",
          "deviceId"
        ],
        "type": "object"
      },
      "DashboardLayoutDto": {
        "additionalProperties": false,
        "properties": {
          "customItems": {
            "items": {
              "$ref": "#/components/schemas/CustomLayoutItemDto"
            },
            "nullable": true,
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "layoutMode": {
            "nullable": true,
            "type": "string"
          },
          "name": {
            "nullable": true,
            "type": "string"
          },
          "tiles": {
            "items": {
              "$ref": "#/components/schemas/DashboardTileDto"
            },
            "nullable": true,
            "type": "array"
          },
          "updatedAtUtc": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "layoutMode",
          "name",
          "tiles",
          "updatedAtUtc"
        ],
        "type": "object"
      },
      "DashboardRainfallDto": {
        "additionalProperties": false,
        "properties": {
          "dailyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "deviceId": {
            "nullable": true,
            "type": "string"
          },
          "deviceName": {
            "nullable": true,
            "type": "string"
          },
          "eventRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "lastRain": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "monthlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "receivedAtUtc": {
            "format": "date-time",
            "type": "string"
          },
          "timestampUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "weeklyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "yearlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          }
        },
        "required": [
          "deviceId",
          "receivedAtUtc"
        ],
        "type": "object"
      },
      "DashboardTileDto": {
        "additionalProperties": false,
        "properties": {
          "deviceId": {
            "nullable": true,
            "type": "string"
          },
          "h": {
            "format": "int32",
            "type": "integer"
          },
          "i": {
            "nullable": true,
            "type": "string"
          },
          "metricKey": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          },
          "w": {
            "format": "int32",
            "type": "integer"
          },
          "x": {
            "format": "int32",
            "type": "integer"
          },
          "y": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "h",
          "i",
          "type",
          "w",
          "x",
          "y"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": {},
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SaveDashboardLayoutCommand": {
        "additionalProperties": false,
        "properties": {
          "customItems": {
            "items": {
              "$ref": "#/components/schemas/CustomLayoutItemDto"
            },
            "nullable": true,
            "type": "array"
          },
          "layoutMode": {
            "nullable": true,
            "type": "string"
          },
          "tiles": {
            "items": {
              "$ref": "#/components/schemas/DashboardTileDto"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "description": "Paste your Auth0 access token (without \u0027Bearer \u0027 prefix).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "BFF API for the Ambient Weather Dashboard web application.",
    "title": "Ambient Weather Dashboard API",
    "version": "v1"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/dashboard/current": {
      "get": {
        "parameters": [
          {
            "description": "Data source: \u0060own\u0060 (default) for the user\u0027s own station;\n\u0060neighbors\u0060 for an aggregated reading from nearby public stations.",
            "in": "query",
            "name": "source",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentReadingDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentReadingDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentReadingDto"
                }
              }
            },
            "description": "Latest reading."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "No credentials/station configured, or neighbor feature unavailable."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limit exceeded."
          },
          "503": {
            "description": "Ambient circuit breaker is open."
          }
        },
        "security": [
          {}
        ],
        "summary": "Returns the latest current reading for the user\u0027s default weather station,\nor an aggregated neighbor reading when source is \u0060neighbors\u0060.",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/dashboard/daily-extremes": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyExtremaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyExtremaDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DailyExtremaDto"
                }
              }
            },
            "description": "Daily temperature extrema for the primary station."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "No Ambient credentials saved or no station configured."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limit exceeded."
          }
        },
        "security": [
          {}
        ],
        "summary": "Returns the daily high/low outdoor and indoor temperatures for the user\u0027s default station,\ncomputed from stored readings for the current UTC calendar day.\nReturns null values for any sensor that has no readings today.",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/dashboard/layout": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardLayoutDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardLayoutDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardLayoutDto"
                }
              }
            },
            "description": "The active layout with all tile configurations."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limit exceeded."
          }
        },
        "security": [
          {}
        ],
        "summary": "Returns the user\u0027s active dashboard layout, seeding a suggested default on first access.",
        "tags": [
          "Dashboard"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SaveDashboardLayoutCommand"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDashboardLayoutCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDashboardLayoutCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardLayoutDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardLayoutDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardLayoutDto"
                }
              }
            },
            "description": "The saved layout echoed back with its persisted identifier and timestamp."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed \u2014 invalid tile structure or unrecognized metric key."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limit exceeded."
          }
        },
        "security": [
          {}
        ],
        "summary": "Saves the user\u0027s dashboard tile layout.\nValidates tile structure, metric keys, and metric reference format before persisting.\nStation IDs in custom items are display configuration and are not validated against ownership.",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/dashboard/rainfall": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardRainfallDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardRainfallDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardRainfallDto"
                }
              }
            },
            "description": "Rainfall accumulation values for the primary station."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "No Ambient credentials saved or no station configured."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limit exceeded."
          },
          "503": {
            "description": "Ambient circuit breaker is open."
          }
        },
        "security": [
          {}
        ],
        "summary": "Returns a rainfall accumulation snapshot for the user\u0027s default weather station.\nReads from the server-side latest-reading cache first; falls back to the Ambient REST API\non cache miss.",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
  },
...
}